# contracts/api.yaml
openapi: 3.0.3
info:
  title: Booking Link API
  version: 0.1.0
paths:
...
  /api/admin/google/start:
    get:
      summary: Start owner Google OAuth flow.
      parameters:
        - name: secret
          in: query
          required: true
          schema:
            type: string
      responses:
        "307":
          description: Redirect to Google OAuth.
...